+++ src/SelfInitializingFakes/Infrastructure/CompoundTypeConverter.cs
@@ -26,21 +26,23 @@ namespace SelfInitializingFakes.Infrastructure
         /// Potentially converts an unserializable object to a more serializable form.
         /// </summary>
         /// <param name="input">An input object.</param>
+        /// <param name="mainConverter">A comprehensive converter that may be used to further convert the output, if required.</param>
         /// <param name="output">An output object. Will be assigned to a simpler representation of <paramref name="input"/>, if this converter knows how.</param>
         /// <returns><c>true</c> if the conversion happened, otherwise <c>false</c>. Good for building a chain of responsibility.</returns>
-        public bool ConvertForRecording(object? input, out object? output) =>
-            this.first.ConvertForRecording(input, out output) ||
-            this.second.ConvertForRecording(input, out output);
+        public bool ConvertForRecording(object? input, ITypeConverter mainConverter, out object? output) =>
+            this.first.ConvertForRecording(input, mainConverter, out output) ||
+            this.second.ConvertForRecording(input, mainConverter, out output);
 
         /// <summary>
         /// Potentially converts the serializable form of an object back to its unserializable form.
         /// </summary>
         /// <param name="deserializedType">The desired deserialized type.</param>
         /// <param name="input">An input object.</param>
+        /// <param name="mainConverter">A comprehensive converter that may be used to further convert the output, if required.</param>
         /// <param name="output">An output object. Will be reconstituted from its simpler representation as <paramref name="input"/>, if this converter knows how.</param>
         /// <returns><c>true</c> if the conversion happened, otherwise <c>false</c>. Good for building a chain of responsibility.</returns>
-        public bool ConvertForPlayback(Type deserializedType, object? input, out object? output) =>
-            this.first.ConvertForPlayback(deserializedType, input, out output) ||
-            this.second.ConvertForPlayback(deserializedType, input, out output);
+        public bool ConvertForPlayback(Type deserializedType, object? input, ITypeConverter mainConverter, out object? output) =>
+            this.first.ConvertForPlayback(deserializedType, input, mainConverter, out output) ||
+            this.second.ConvertForPlayback(deserializedType, input, mainConverter, out output);
     }
 }