+++ src/SelfInitializingFakes/Infrastructure/LazyTypeConverter.cs
@@ -15,9 +15,10 @@ namespace SelfInitializingFakes.Infrastructure
         /// Potentially converts an unserializable object to a more serializable form.
         /// </summary>
         /// <param name="input">An input object.</param>
+        /// <param name="mainConverter">A comprehensive converter that may be used to further convert the output, if required.</param>
         /// <param name="output">An output object. Will be assigned to a simpler representation of <paramref name="input"/>, if this converter knows how.</param>
         /// <returns><c>true</c> if the conversion happened, otherwise <c>false</c>. Good for building a chain of responsibility.</returns>
-        public bool ConvertForRecording(object? input, out object? output)
+        public bool ConvertForRecording(object? input, ITypeConverter mainConverter, out object? output)
         {
             output = null;
             if (input is null)
@@ -29,6 +30,11 @@ namespace SelfInitializingFakes.Infrastructure
             if (inputType.IsInstanceOf(typeof(Lazy<>)))
             {
                 output = inputType.GetProperty("Value").GetGetMethod().Invoke(input, Type.EmptyTypes);
+                if (mainConverter.ConvertForRecording(output, mainConverter, out object? furtherConvertedOutput))
+                {
+                    output = furtherConvertedOutput;
+                }
+
                 return true;
             }
 
@@ -40,16 +46,28 @@ namespace SelfInitializingFakes.Infrastructure
         /// </summary>
         /// <param name="deserializedType">The desired deserialized type.</param>
         /// <param name="input">An input object.</param>
+        /// <param name="mainConverter">A comprehensive converter that may be used to further convert the output, if required.</param>
         /// <param name="output">An output object. Will be reconstituted from its simpler representation as <paramref name="input"/>, if this converter knows how.</param>
         /// <returns><c>true</c> if the conversion happened, otherwise <c>false</c>. Good for building a chain of responsibility.</returns>
-        public bool ConvertForPlayback(Type deserializedType, object? input, out object? output)
+        public bool ConvertForPlayback(Type deserializedType, object? input, ITypeConverter mainConverter, out object? output)
         {
             if (deserializedType.IsInstanceOf(typeof(Lazy<>)))
             {
                 var typeOfLazyResult = deserializedType.GetGenericArguments()[0];
-                var method = CreateLazyGenericDefinition.MakeGenericMethod(typeOfLazyResult);
-                output = method.Invoke(null, new object?[] { input });
-                return true;
+
+                if (input is null || input.GetType() == typeOfLazyResult)
+                {
+                    var method = CreateLazyGenericDefinition.MakeGenericMethod(typeOfLazyResult);
+                    output = method.Invoke(null, new object?[] { input });
+                    return true;
+                }
+
+                if (mainConverter.ConvertForPlayback(typeOfLazyResult, input, mainConverter, out object? convertedInput))
+                {
+                    var method = CreateLazyGenericDefinition.MakeGenericMethod(typeOfLazyResult);
+                    output = method.Invoke(null, new object?[] { convertedInput });
+                    return true;
+                }
             }
 
             output = null;