+++ tests/Acceptance/Helpers/ISampleService.cs
@@ -23,5 +23,7 @@ namespace SelfInitializingFakes.Tests.Acceptance.Helpers
         Task<int> TaskIntReturningMethod();
 
         Lazy<Task<int>> LazyTaskIntReturningMethod();
+
+        Task<Lazy<int>> TaskLazyIntReturningMethod();
     }
 }